donation chapter incorporated into 'rest of public bookdown'
@DMoss123 @peterhurford  @oskasf

Rather than being in a separate place, the Donations chapter is now integrated with the other public parts of the bookdown (and it should show up [here](https://rethinkpriorities.github.io/ea_data_public/eas-donations.html), I think.

I will still maintain that other version that was "just the donations chapter" for a while, as we had hypothes.is comments embedded in it and on that's the link I share d on Teams.

But this (https://rethinkpriorities.github.io/ea_data_public/eas-donations.html) should be the link we share publlicly and link to the forum post.