add 'new_style_version_pre_bootstrap' output folder
This uses the styling that comes just before the bootstrap template that Willem built into the RP R package.

- This styling has some great features IMO (e.g., foldables have a title)
- but has a few imperfections; especially expanding 'footnotes' can get cut off on the left

In terms of 'making the donation post concise and year-independent', this is about 90% done in that direction (but we also have some judgement calls to make about what stuff to include 'regularly', 'one-off', etc)

Please jump right to chapter 3 on donations.

I